+++ blocks/src/contants.js
@@ -1,0 +1,8 @@
+
+export const ALL_BLOCKS = [
+    'lichtberg/group',
+    'lichtberg/image',
+    'lichtberg/link',
+    'lichtberg/shortcode',
+    'lichtberg/text'
+];