+++ blocks/src/section/edit.js
@@ -1,8 +1,12 @@
 import { __ } from '@wordpress/i18n';
-import { useBlockProps, InnerBlocks, InspectorControls, useInnerBlocksProps } from '@wordpress/block-editor';
-import { PanelBody, SelectControl } from '@wordpress/components';
-import { ALL_BLOCKS, StylePanel, generateClasses } from '../lichtbergHelpers';
+import { useBlockProps, InspectorControls, useInnerBlocksProps } from '@wordpress/block-editor';
+import { PanelBody } from '@wordpress/components';
 import './editor.scss';
+
+import { ALL_BLOCKS } from '../contants';
+import StylePanel from '../components/StylePanel';
+import generateClasses from '../helpers/generateClasses';
+
 
 export default function Edit({ attributes, setAttributes }) {
 	const blockProps = useBlockProps({